+++ packages/eyes-common/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [unversioned]
+
+
+### Features
+
+* **eyes-common:** Add dontCloseBatches property to Configuration. [Trello_1238](https://trello.com/c/5Hy0nGkN)
+
 ## [3.12.2](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-common@3.12.1...@applitools/eyes-common@3.12.2) (2019-11-28)